refactor(PAT-62): Modularize, creating TitleEditable.
