+++ com/frontend/app/src/pages/BoardPage.jsx
@@ -4,21 +4,20 @@ import Dashboard from '../containers/Dashboard';
 import Board from '../containers/Board';
 import { useParams } from 'react-router-dom';
 import getDataBoard, { updatBoardTitle } from '../utils/board';
-import Title from '../components/Title';
-import { TextField } from '@mui/material';
+import TitleEditable from '../components/TitleEditable';
 
-const titleStyle = {
-    style: {
-      fontSize: '1.5rem', // Adjust the font size as needed
-      color: 'white',    // Text color
-    },
-}
+
 const theme = createTheme();
+
+
+/**
+ * BoardPage is a React component representing a page displaying the content of a specific board.
+ *
+ * @returns {JSX.Element} - A JSX element representing the board page.
+ */
 export default function BoardPage( ) {
     const { id_board } = useParams();
     const [title, setTitle] = React.useState('Undefine Title');
-    const [isEditing, setIsEditing] = React.useState(false);
-    const [editedTitle, setEditedTitle] = React.useState(title);
 
     useEffect(() => {
         const fetchBoardTitle = async () => {
@@ -33,40 +32,18 @@ export default function BoardPage( ) {
         };
         fetchBoardTitle();
     }, [id_board]);
-
-
-    const handleDoubleClick = () => {
-            setEditedTitle(title);
-            setIsEditing(true);
-    };
-
-    const handleTitleChange = (event) => {
-        setEditedTitle(event.target.value);
-    };
-
-    const handleTitleBlur = async () => {
-        const newTitle = await updatBoardTitle(id_board, editedTitle);
-        if (newTitle !== null && newTitle !== undefined) {
-            setTitle(newTitle);
+    
+    const saveTitle = async (newTitle) => {
+        const responseTitle = await updatBoardTitle(id_board, newTitle);
+        if (responseTitle !== null && responseTitle !== undefined) {
+            setTitle(responseTitle);
         }
-        setIsEditing(false);
     };
+    const titleComponent = TitleEditable({title, saveTitle});
 
     return (
         <ThemeProvider theme={theme}>
-            <Dashboard title={isEditing ? (
-            <TextField
-                value={editedTitle}
-                onChange={handleTitleChange}
-                onBlur={handleTitleBlur}
-                autoFocus
-                fullWidth
-                variant="outlined"
-                InputProps={titleStyle}
-                />
-            ) : (
-                <Title onDoubleClick={handleDoubleClick}>{title}</Title>
-            )}>
+            <Dashboard title={titleComponent}>
                 <Board  id={String(id_board)} />
             </Dashboard>
         </ThemeProvider>
